Sorry for the wait. Here's what I changed:
Changed: index.jsp, SignInServlet, GuestServlet, signInController
Added: highScoreController

SignIn will now push through to Index if the username and password are
correct.
Guess will now push through to Index if the username is valid.
Both will display the username at the top of the Index page.

The Index page will show the high scores which are a part of a fake
database.
They will most likely not work for you until we get the Derby project
into Git.

***Note: My version of JDK is different, but since Eclipse can easily
change which version the project is using, I didn't think much of it.
Let me know if we're sticking to one version so I can grab that.
Let me know if you have any questions.