Fixes Before Final Presentation
Fixed a bug with the Minesweeper game where some tiles would not
correctly display the right number.
Added missing error messages to the sign in and sign up pages.
Edited the title of the index page to "Home".
Changed the color of the win/lose and final score messages to make them
readable against the background.
Added a few comments here and there.